pyproject.toml
[tool.poetry]
name = "dash-app-dbt-reactflow"
version = "0.1.0"
description = ""
authors = ["Josh Peak <neozenith.dev@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "dash_app_dbt_reactflow"}]

[tool.poetry.dependencies]
python = "^3.10"
dash = "^2.13.0"
dash-bootstrap-components = "^1.4.2"
pandas = "^2.1.0"
fastapi = "^0.103.1"
hypercorn = "^0.14.4"
dbt-core = "^1.6.2"
dbt-postgres = "^1.6.2"


[tool.poetry.group.dev.dependencies]
invoke = "^2.2.0"
black = "^23.7.0"
flake8 = "^6.1.0"
mypy = "^1.5.1"
md-toc = "^8.2.0"
pytest = "^7.4.1"
coverage = "^7.3.0"
pytest-cov = "^4.1.0"
pytest-flake8 = "^1.1.1"
flake8-docstrings = "^1.7.0"
isort = "^5.12.0"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
multi_line_output = 3
import_heading_stdlib = "Standard Library"
import_heading_firstparty = "Our Libraries"
import_heading_thirdparty = "Third Party"

[tool.mypy]
exclude = ["tests/", "tasks\\.py"]
pretty = true
show_error_codes = true
show_column_numbers = true
show_error_context = true
ignore_missing_imports = true
follow_imports = "silent"
disallow_incomplete_defs = true
disallow_untyped_defs = false
strict = false

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-s -vvv --color=yes --cov=. --no-cov-on-fail"

[tool.coverage.run]
branch = true
omit = ["tests/*", "**/__init__.py", "tasks.py"]
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
